auction creation works
