+++ core/src/calleeFunctions/rETHCurveUniv3Callee.ts
@@ -1,11 +1,11 @@
 import type { CalleeFunctions, CollateralConfig } from '../types';
 import { ethers } from 'ethers';
 import BigNumber from '../bignumber';
-import { ETH_NUMBER_OF_DIGITS } from '../constants/UNITS';
-import getContract, { getContractAddressByName, getJoinNameByCollateralType } from '../contracts';
+import { getContractAddressByName, getJoinNameByCollateralType } from '../contracts';
 import { encodeRoute, convertCollateralToDai } from './helpers/uniswapV3';
 import { convertStethToEth } from './helpers/curve';
 import { convertRethToWsteth } from './helpers/rocket';
+import { convertWstethToSteth } from './helpers/wsteth';
 
 export const CHARTER_MANAGER_ADDRESS = '0x8377CD01a5834a6EaD3b7efb482f678f2092b77e';
 
@@ -32,17 +32,14 @@ const getCalleeData = async function (
 
 const getMarketPrice = async function (
     network: string,
-    collateral: CollateralConfig,
+    _collateral: CollateralConfig,
     collateralAmount: BigNumber
 ): Promise<BigNumber> {
     // convert rETH into wstETH
     const wstethAmount = await convertRethToWsteth(network, collateralAmount);
 
     // convert wstETH into stETH (unwrap)
-    const collateralContract = await getContract(network, collateral.symbol);
-    const collateralIntegerAmount = wstethAmount.shiftedBy(collateral.decimals).toFixed(0);
-    const stethIntegerAmount = await collateralContract.getStETHByWstETH(collateralIntegerAmount);
-    const stethAmount = new BigNumber(stethIntegerAmount._hex).shiftedBy(-ETH_NUMBER_OF_DIGITS);
+    const stethAmount = await convertWstethToSteth(network, wstethAmount);
 
     // convert stETH into ETH
     const ethAmount = await convertStethToEth(network, stethAmount);
